# clients/users-api/specs/api.yaml
openapi: 3.0.3
info:
  title: users service api
  version: 1.0.0
  description: API для сервиса users

paths:
  /auth/register:
    post:
      description: Регистрация пользователя
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#./definitions.yaml#/components/schemas/UserRegistrationRequest'
      responses:
        '200':
          description: Ответ сервера с данными о регистрации
          content:
            application/json:
              schema:
                $ref: '#./definitions.yaml#/components/schemas/UserRegistrationResponse'

  /auth/login:
    post:
      description: Вход в систему
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#./definitions.yaml#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: Ответ сервера с данными о регистрации
          content:
            application/json:
              schema:
                $ref: '#./definitions.yaml#/components/schemas/UserLoginResponse'

  /auth/change-password:
    post:
      description: Смена пароля пользователя
      tags:
        - auth
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: './definitions.yaml#/components/parameters/TokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#./definitions.yaml#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Результат смены пароля
          content:
            application/json:
              schema:
                $ref: '#./definitions.yaml#/components/schemas/ChangePasswordResponse'

  /auth/logout:
    post:
      description: Выход из аккаунта
      tags:
        - auth
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Информация о выходе из аккаунта (если ок то пустое тело)
          content:
            application/json:
              schema:
                $ref: '#./definitions.yaml#/components/schemas/LogoutResponse'

  /users/get-user-credentials:
    get:
      parameters:
        - $ref: './definitions.yaml#/components/parameters/TokenHeader'
      description: Получение credentials пользователя по токену
      tags:
        - user
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: '#./definitions.yaml#/components/schemas/GetUserCredentialsResponse'


  /token/public-jwt:
    get:
      description: Получает публичный ключ для шифрования jwt токенов
      tags:
        - token
      responses:
        '200':
          description: Публичный jwt ключ
          content:
            application/json:
              schema:
                $ref: '#./definitions.yaml#/components/schemas/PublicJwtResponse'

  /token/refresh:
    post:
      description: Обновление access-токена по refresh-токену
      tags:
        - token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#./definitions.yaml#/components/schemas/RefreshToken'
      responses:
        '200':
          description: Новый access token
          content:
            application/json:
              schema:
                $ref: '#./definitions.yaml#/components/schemas/RefreshTokenResponse'

  /users/device-token:
    put:
      description: Сохранить FCM-токен устройства текущего пользователя
      tags:
        - user
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: './definitions.yaml#/components/parameters/TokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './definitions.yaml#/components/schemas/DeviceTokenRequest'
      responses:
        '200':
          description: Токен сохранён

  /users/internal/device-token:
    post:
      description: Получить FCM-токен пользователя по userId (только для сервисов)
      tags:
        - user
      security:
        - apiKeyAuth: [ ]
      parameters:
        - $ref: '../definitions.yaml#/components/parameters/ApiKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './definitions.yaml#/components/schemas/GetFcmTokenRequest'
      responses:
        '200':
          description: FCM-токен пользователя
          content:
            application/json:
              schema:
                $ref: './definitions.yaml#/components/schemas/FcmTokenResponse'

  /users/get-by-id:
    post:
      description: Получение credentials пользователя по ID (поддерживает пользовательский JWT или сервисный API-ключ)
      tags:
        - user
      security:
        - bearerAuth: [ ]
        - apiKeyAuth: [ ]
      parameters:
        - $ref: '../definitions.yaml#/components/parameters/ApiKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../definitions.yaml#/components/schemas/UserByIdRequest'
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: '../definitions.yaml#/components/schemas/GetUserCredentialsResponse'

  /users/profile:
    delete:
      description: Удаление профиля текущего пользователя (soft delete)
      tags:
        - user
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: './definitions.yaml#/components/parameters/TokenHeader'
      responses:
        '200':
          description: Результат удаления профиля
          content:
            application/json:
              schema:
                $ref: './definitions.yaml#/components/schemas/DeleteProfileResponse'

  /users/email:
    patch:
      description: Смена email текущего пользователя
      tags:
        - user
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: './definitions.yaml#/components/parameters/TokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './definitions.yaml#/components/schemas/ChangeEmailRequest'
      responses:
        '200':
          description: Результат смены email
          content:
            application/json:
              schema:
                $ref: './definitions.yaml#/components/schemas/ChangeEmailResponse'

  /auth/verification-code:
    post:
      description: Отправить код подтверждения на email
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './definitions.yaml#/components/schemas/SendVerificationCodeRequest'
      responses:
        '200':
          description: Результат отправки кода
          content:
            application/json:
              schema:
                $ref: './definitions.yaml#/components/schemas/SendVerificationCodeResponse'

  /auth/check-verification-code:
    post:
      description: Проверить введённый код подтверждения
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './definitions.yaml#/components/schemas/CheckVerificationCodeRequest'
      responses:
        '200':
          description: Результат проверки кода
          content:
            application/json:
              schema:
                $ref: './definitions.yaml#/components/schemas/CheckVerificationCodeResponse'

  /auth/reset-password:
    post:
      description: Сбросить пароль с помощью кода подтверждения
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './definitions.yaml#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Результат сброса пароля
          content:
            application/json:
              schema:
                $ref: './definitions.yaml#/components/schemas/ResetPasswordResponse'